finish refactoring rejectCode to msgRejectCode
