+++ examples/stable/test/property/ic_api/msg_reject_code/.gitignore
@@ -1,0 +1,4 @@
+.azle
+.dfx
+dfx_generated
+node_modules
