+++ examples/stable/test/property/ic_api/msg_reject_code/cuzz.json
@@ -1,0 +1,3 @@
+{
+    "expectedErrors": ["Thrown error", "Rejection with message"]
+}
