+++ examples/stable/test/property/ic_api/msg_reject_code/src/rejector/index.ts
@@ -1,0 +1,16 @@
+import { IDL, msgReject, query } from 'azle';
+
+export default class {
+    @query([], IDL.Text)
+    throwError(): string {
+        throw new Error('Thrown error');
+    }
+
+    @query([], IDL.Text, { manual: true })
+    rejectWithMessage(): void {
+        msgReject('Rejection with message');
+    }
+
+    @query
+    noError(): void {}
+}
