+++ examples/stable/test/property/ic_api/msg_reject_code/test/test.ts
@@ -1,0 +1,5 @@
+import { runTests } from 'azle/test';
+
+import { getTests } from './tests';
+
+runTests(getTests());
