--- examples/stable/test/property/ic_api/reject_code/.gitignore
@@ -1,4 +1,0 @@
-.azle
-.dfx
-dfx_generated
-node_modules
