--- examples/stable/test/property/ic_api/reject_code/cuzz.json
@@ -1,3 +1,0 @@
-{
-    "expectedErrors": ["Thrown error", "Rejection with message"]
-}
