--- examples/stable/test/property/ic_api/reject_code/jest.config.js
@@ -1,11 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    testTimeout: 100_000_000,
-    transform: {
-        '^.+\\.ts$': ['ts-jest', { isolatedModules: true }],
-        '^.+\\.js$': 'ts-jest'
-    },
-    transformIgnorePatterns: ['/node_modules/(?!(azle)/)'] // Make sure azle is transformed
-};
