--- examples/stable/test/property/ic_api/reject_code/package.json
@@ -1,15 +1,0 @@
-{
-    "scripts": {
-        "pretest": "tsx test/pretest.ts",
-        "test": "jest"
-    },
-    "dependencies": {
-        "azle": "0.25.0"
-    },
-    "devDependencies": {
-        "jest": "^29.7.0",
-        "ts-jest": "^29.1.4",
-        "tsx": "^4.15.7",
-        "typescript": "^5.2.2"
-    }
-}
