--- examples/stable/test/property/ic_api/reject_code/test/test.ts
@@ -1,5 +1,0 @@
-import { runTests } from 'azle/test';
-
-import { getTests } from './tests';
-
-runTests(getTests());
