--- examples/stable/test/property/ic_api/reject_code/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "strict": true,
-        "target": "ES2020",
-        "module": "ES2020",
-        "moduleResolution": "node",
-        "allowJs": true,
-        "outDir": "HACK_BECAUSE_OF_ALLOW_JS"
-    }
-}
