+++ src/lib/stable/ic_apis/index.ts
@@ -18,11 +18,11 @@ export { msgCyclesAvailable } from './msg_cycles_available';
 export { msgCyclesRefunded } from './msg_cycles_refunded';
 export { msgMethodName } from './msg_method_name';
 export { msgReject } from './msg_reject';
+export { msgRejectCode } from './msg_reject_code';
 export { msgRejectMsg } from './msg_reject_msg';
 export { msgReply } from './msg_reply';
 export { notify } from './notify';
 export { performanceCounter } from './performance_counter';
-export { rejectCode, RejectionCode } from './reject_code';
 export { setCertifiedData } from './set_certified_data';
 export { setTimer } from './set_timer';
 export { setTimerInterval } from './set_timer_interval';
